added degree

+++ permgroup/_clean_.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
-rm *.c
-rm *.pyc
-rm *.so
+WORK_DIR="$(pwd)"
 
-rm -rf build
+rm -f *.c
+rm -f *.pyc
+rm -f *.so
+
+if [[ -d "$WORK_DIR/build" ]]; then
+    rm -rf build
+    echo "- removed build directory"
+fi
